# server/openapi.yml
openapi: 3.0.0
servers:
    - description: SwaggerHub API Auto Mocking
      url: https://virtserver.swaggerhub.com/C4theBomb/qOverflow/1.0.0
    - description: Localhost
      url: http://localhost:3001
info:
    version: '1.0.0'
    title: qOverflow API
    description: >-
        This API will interact with the BDPA API to expand operations and cache data to bypass rate limitations. It also implements enforcement features to prevent unauthorized operations. Authorization is done through the `authorization` header and has two scopes, `bearer` and `basic`.
paths:
...
    /users/{username}:
        parameters:
            - name: username
              in: path
              required: true
              schema:
                  type: string
                  example: 'pinapplezzz'
              description: >-
                  This should be the username of the desired user.
        get:
            summary: Get a user by their username
            responses:
                '200':
                    description: OK
                    content:
                        'application/json':
                            schema:
                                type: object
                                properties:
                                    user:
                                        $ref: '#/components/schemas/User'
                '500':
                    $ref: '#/components/responses/ServerError'
            tags:
                - users
...
    /questions/{question_id}:
        parameters:
            - $ref: '#/components/parameters/question_id'
        get:
            summary: Get a question by its question_id
            responses:
                '200':
                    description: OK
                    content:
                        'application/json':
                            schema:
                                type: object
                                properties:
                                    question:
                                        $ref: '#/components/schemas/Question'
                '500':
                    $ref: '#/components/responses/ServerError'
            tags:
                - questions
        patch:
            summary: Modify a question by its question_id
            security:
                - BearerAuth: []
            requestBody:
                content:
                    'application/json':
                        schema:
                            type: object
                            properties:
                                text:
                                    type: string
            responses:
                '200':
                    description: OK
                '400':
                    $ref: '#/components/responses/BadRequest'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '500':
                    $ref: '#/components/responses/BadRequest'
            tags:
                - questions
...
    /questions/{question_id}/comments:
        parameters:
            - $ref: '#/components/parameters/question_id'
        get:
            summary: Get all comments of a question
            responses:
                '200':
                    description: OK
                    content:
                        'application/json':
                            schema:
                                type: object
                                properties:
                                    comments:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/Comment'
                '500':
                    $ref: '#/components/responses/ServerError'
            tags:
                - comments
        post:
            summary: Create a comment on a question
            security:
                - BearerAuth: []
            requestBody:
                content:
                    'application/json':
                        schema:
                            type: object
                            properties:
                                text:
                                    type: string
                                    maxLength: 150
            responses:
                '200':
                    description: OK
                '400':
                    $ref: '#/components/responses/BadRequest'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '500':
                    $ref: '#/components/responses/BadRequest'
            tags:
                - comments
    /questions/{question_id}/comments/{comment_id}:
        parameters:
            - $ref: '#/components/parameters/question_id'
            - $ref: '#/components/parameters/comment_id'
        delete:
            summary: Delete a question comment
            security:
                - BearerAuth: []
            responses:
                '200':
                    description: OK
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '500':
                    $ref: '#/components/responses/ServerError'
            tags:
                - comments
...
    /questions/{question_id}/answers/{answer_id}:
        parameters:
            - $ref: '#/components/parameters/question_id'
            - $ref: '#/components/parameters/answer_id'
        patch:
            summary: Modify an answer by its answer_id
            security:
                - BearerAuth: []
            requestBody:
                content:
                    'application/json':
                        schema:
                            type: object
                            properties:
                                text:
                                    type: string
            responses:
                '200':
                    description: OK
                    content:
                        'application/json':
                            schema:
                                type: object
                                properties:
                                    answer:
                                        $ref: '#/components/schemas/Answer'
                '400':
                    $ref: '#/components/responses/BadRequest'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '500':
                    $ref: '#/components/responses/BadRequest'
            tags:
                - answers
...
components:
    securitySchemes:
        BasicAuth:
            type: http
            scheme: basic
            description: >-
                This is the username and password authentication method used for the login route. The authorization header should be equal to `basic {base64 encoded username:password}` encoded into base64. The base64 encoding can be found by running `btoa(username:password)`.
        BearerAuth:
            type: http
            scheme: bearer
            bearerFormat: jwt
            description: >-
                This is the JWT token authentication method used for other routes requiring user privileges. The authorization header should be equal to `bearer {JTW Token}`. The token is returned in the token property of the `/users/login` response.
    schemas:
        Answer:
            type: object
            properties:
                id:
                    type: string
                    example: '62d8727d6c584b3b12eb23be'
                question_id:
                    type: string
                    example: '62d5c7ad1b8e0da8804bde22'
                creator:
                    type: string
                    example: 'pinapplezzz'
                text:
                    type: string
                upvotes:
                    type: integer
                downvotes:
                    type: integer
                accepted:
                    type: boolean
                createdAt:
                    type: string
                    format: date-time
        Comment:
            type: object
            properties:
                id:
                    type: string
                    example: '62d87292afa7f2e7e81f1470'
                parent_id:
                    type: string
                    example: '62d5c7ad1b8e0da8804bde22'
                creator:
                    type: string
                    example: 'pinapplezzz'
                text:
                    type: string
                upvotes:
                    type: integer
                downvotes:
                    type: integer
                createdAt:
                    type: string
                    format: date-time
...
        Question:
            type: object
            properties:
                id:
                    type: string
                    example: '62d872b85acd7bccf9087d77'
                title:
                    type: string
                    maxLength: 150
                text:
                    type: string
                    maxLength: 3000
                status:
                    type: string
                    enum: ['open', 'protected', 'closed']
                views:
                    type: integer
                answers:
                    type: integer
                comments:
                    type: integer
                upvotes:
                    type: integer
                downvotes:
                    type: integer
                hasAccepted:
                    type: boolean
                creator:
                    type: string
                    example: 'pinapplezzz'
                reopen:
                    type: array
                    items:
                        type: string
                        example: 'pinapplezzz'
                close:
                    type: array
                    items:
                        type: string
                        example: 'pinapplezzz'
                protect:
                    type: array
                    items:
                        type: string
                        example: 'pinapplezzz'
                createdAt:
                    type: string
                    format: date-time
        User:
            type: object
            properties:
                username:
                    type: string
                    example: 'pinapplezzz'
                email:
                    type: string
                    format: email
                points:
                    type: integer
...
    responses:
        BadRequest:
            description: Bad request
            content:
                'application/json':
                    schema:
                        type: object
                        properties:
                            error:
                                type: string
                                example: 'Your request is missing information.'
        ServerError:
            description: Server Error
            content:
                'application/json':
                    schema:
                        type: object
                        properties:
                            error:
                                type: string
                                example: 'Oops! Something went wrong.'
...
        Forbidden:
            description: User not allowed
            content:
                'application/json':
                    schema:
                        type: object
                        properties:
                            error:
                                type: string
                                example: 'You are not authorized to do that.'
        NotFound:
            description: Not Found
            content:
                'application/json':
                    schema:
                        type: object
                        properties:
                            error:
                                type: string
                                example: 'That resource was not found. Please check your request and try again.'
    parameters:
        question_id:
            name: question_id
            in: path
            required: true
            schema:
                type: string
                example: '62d61d428849dda1c15a8b64'
            description: >-
                This should be the object id of the desired question.
        answer_id:
            name: answer_id
            in: path
            required: true
            schema:
                type: string
                example: '62d61ffd5f20e06f3cb350c7'
            description: >-
                This should be the object id of the desired answer.
        comment_id:
            name: comment_id
            in: path
            required: true
            schema:
                type: string
                example: '62d620175bdb35fbb86e33cb'
            description: >-
                This should be the object id of the desired comment.